Check if /opt is owned by root before attempting to install

+++ scripts/firefox-nightly/update-nightly.sh
@@ -5,9 +5,13 @@ TARBALL="firefox-$RELEASE.en-US.linux-x86_64.tar.bz2"
 
 wget https://ftp.mozilla.org/pub/firefox/nightly/latest-mozilla-central/$TARBALL
 
+if [ $(stat -c %U /opt) == "root" ]; then
+   echo "/opt/ is not available for writing."
+   exit
+fi
 if [ -d "/opt/firefox" ]; then
-  rm -r /opt/firefox
-  mkdir /opt/firefox
+   rm -r /opt/firefox
+   mkdir /opt/firefox
 fi
 
 tar -xvjf $TARBALL -C /opt
